+++ lib/order_query/sql/order_by.rb
@@ -13,47 +13,51 @@ module OrderQuery
 
       # @return [String]
       def build_reverse
-        @reverse_sql ||= join_order_by_clauses order_by_reverse_sql_clauses
+        @reverse_sql ||= join_order_by_clauses order_by_sql_clauses(true)
       end
 
       protected
 
       # @return [Array<String>]
-      def order_by_sql_clauses
-        @columns.map { |cond| column_clause cond }
+      def order_by_sql_clauses(reverse = false)
+        @columns.map { |col| column_clause col, reverse }
       end
 
-      def column_clause(cond)
-        dir_sql = sort_direction_sql cond.order
-        col_sql = cond.column_name
-        if cond.order_enum
-          cond.order_enum.map { |v| "#{col_sql}=#{cond.quote v} #{dir_sql}" }.join(', ').freeze
+      def column_clause(col, reverse = false)
+        if col.order_enum
+          column_clause_enum col, reverse
         else
-          "#{col_sql} #{dir_sql}".freeze
+          column_clause_ray col, reverse
         end
       end
 
-      SORT_DIRECTIONS = [:asc, :desc].freeze
+      def column_clause_ray(col, reverse = false)
+        "#{col.column_name} #{sort_direction_sql(col, reverse)}".freeze
+      end
+
+      def column_clause_enum(col, reverse = false)
+        enum = col.order_enum
+        # Collapse boolean enum to `ORDER BY column ASC|DESC`
+        if enum == [false, true] || enum == [true, false]
+          return column_clause_ray col, reverse ^ enum.last
+        end
+        enum.map { |v|
+          "#{order_by_value_sql col, v} #{sort_direction_sql(col, reverse)}"
+        }.join(', ').freeze
+      end
+
+      def order_by_value_sql(col, v)
+        "#{col.column_name}=#{col.quote v}"
+      end
+
       # @return [String]
-      def sort_direction_sql(direction)
-        if SORT_DIRECTIONS.include?(direction)
-          direction.to_s.upcase.freeze
-        else
-          raise ArgumentError.new("sort direction must be in #{SORT_DIRECTIONS.map(&:inspect).join(', ')}, is #{direction.inspect}")
-        end
+      def sort_direction_sql(col, reverse = false)
+        col.direction(reverse).to_s.upcase.freeze
       end
 
       # @param [Array<String>] clauses
       def join_order_by_clauses(clauses)
         clauses.join(', ').freeze
-      end
-
-      # @return [Array<String>]
-      def order_by_reverse_sql_clauses
-        swap = {'DESC' => 'ASC', 'ASC' => 'DESC'}
-        order_by_sql_clauses.map { |s|
-          s.gsub(/DESC|ASC/) { |m| swap[m] }
-        }
       end
     end
   end